feat: improve printing to stdout/stderr
Dedicated macros now make a clear distinction for printing. They essentially deliniate everything, and avoid cognitive overhead in remembering when to use println! or eprintln!.
Further, stderr is used to convey any information that is not strictly related to the lilypond notes output, and thus should not be handled by clients for async input. On stderr, we can both share general information, but also errors. Two dedicated macros further add prefixes to the output, so that clients can proprely parse and distinguish between actual errors and simple information.