feat!: Different pedal modes
Another mode which does the opposite of the already existing pedal mode. The enum and strings were renamed to make a clearer distinction, and in case more pedal modes are needed in the future (i.e. different functionality based on which specific pedal was pressed).

BREAKING CHANGE: pedal mode was renamed