refactor: last_chord takes a set of u8 instead of string
A string is not the most robust representation. If a user provides chords to be set as the last_chord, then the order of the notes will be important, even if they are considered the same chord by LilyPond. Using the set, we can effectively avoid the issue where ordering of notes is taken into account.