feat: set previous chord
This is a feature targeted at editor integrations. It is possible that users jump around the file when writing their score. However, the internally stored "last chord" is likely not the right one anymore.
With this functionality, editors can search for the previous chord used in the LilyPond file, and then set it accordingly here. If the real last chord in the LilyPond file is repeated, it will properly return `q`.
This also addresses an issue where a `q` might be inserted wrongly, because the tool had the last chord set incorrectly (due to the user playing notes on the keyboard, before inserting them).