feat: Trailing `+` or `-` to adjust octave in alterations
An alteration of `0:bis` will always cause the note to be one octave too high when pressing a C.

A trailing `+` will cause the note to be placed an octave **higher**.
A trailing `-` will cause the note to be placed an octave **lower**.
You can have multiple `+` or `-` to alter the octave. I.e. a trailing `++` will place the note two octaves higher. A trailing `---` will place it three octaves lower.

In our example, we can thus resolve the issue now by providing the alteration as `0:bis-`.