feat: Modify note generation to also compute relative octave
In relative mode, we need to know the previous note in order to compute
the relative offset for the current note. At every note generation, we
therefore keep track of the absolute MIDI note value, and use that in
the following note generation to only output the relative octave. Within
relative octave entry, the current note therefore always needs to be
added to the `previous_absolute_note_reference` key for the next note to
be computed.

This also means that if we do not know the previous note's absolute
octave, i.e. because it is the first note that is being entered, we
should take advantage of Lilypond's octave checks and output the note
without any octave marks. The octave check will ensure the output
document will have the note in the right octave, and the user will be
prompted to manually fix the octave mark by the lilypond compiler if
needed.

A last consideration are the (global) alterations that we allow,
especially considering we allow entering abritrary text for those. In
both cases, to properly calculate the relative octave, we would need the
user to somehow additionally specify what MIDI note value the given
alteration corresponds to. For local alterations, we would need to have
the user specify the offset from the given alteration to the base note
without alterations. This was deemed not so important for the time
being, so the simple solution is to simply include the absolute octave
as an octave check when outputting an alteration in relative octave
entry.