feat: Allow printing note with an optional octave check
The main purpose is for relative octave entry, where a note for which we
do not know the relative position of (anymore) can have a octave check
added. This way, at least the output score will have the correct octave,
and the user will be prompted by the lilypond compiler to fix the octave
marks if incorrect.