feat: `note_name` renamed to `render` with new LilyNoteRendered return type
This refactor allows us to capture the MIDI note value of the note when
we remove the accidental. The main motivation behind this is due to how
Lilypond calculates the octave:

> If no octave changing mark is used on a pitch, its octave is
> calculated so that the interval with the previous note is less than
> a fifth. This interval is determined without considering accidentals. 

As a result of the last portion, we must also keep track of the MIDI
note value when the accidental is removed from the note. This allows
proper calculation of the octave mark for relative octave entry.

See: https://lilypond.org/doc/v2.25/Documentation/notation/writing-pitches#relative-octave-entry